Update Bill and Item classes and services
Added the issueDate field to the Bill class for database storage and inclusion in the bill pdf.
Changed the variable type of price and total from Integer to Double in the ItemDto, Item, and Bill classes.
Updated prepareData in the BillServiceImpl class and added setDateForm method.
Made necessary adjustments in the saveBill method (in BillServiceImpl) and in the getItemFromMap method (in ItemServiceImpl) to accommodate the changes in variable types.

+++ restaurant-backend-spring/src/main/java/com/ay/restaurant/dto/ItemDto.java
@@ -12,13 +12,13 @@ public class ItemDto {
     private Integer id;
     private String name;
     private String description;
-    private Integer price;
+    private Double price;
     private String status;
     private Integer categoryId;
     private String categoryName;
     private byte[] image;
 
-    public ItemDto(Integer id, String name, String description, Integer price, String status, Integer categoryId, String categoryName) {
+    public ItemDto(Integer id, String name, String description, Double price, String status, Integer categoryId, String categoryName) {
         this.id = id;
         this.name = name;
         this.description = description;
